ab_riscv_interpreter::rv64

Trait Rv64SystemInstructionHandler

pub trait Rv64SystemInstructionHandler<Reg, Memory, PC, CustomError>
where
    Reg: Register<Type = u64>,
    [(); Reg::N]:,
{
    // Required method
    fn handle_ecall(
        &mut self,
        regs: &mut Registers<Reg>,
        memory: &mut Memory,
        program_counter: &mut PC,
    ) -> Result<ControlFlow<()>, ExecutionError<Reg::Type, Rv64Instruction<Reg>, CustomError>>;

    // Provided method
    fn handle_ebreak(
        &mut self,
        _regs: &mut Registers<Reg>,
        _memory: &mut Memory,
        _pc: Reg::Type,
        _instruction: Rv64Instruction<Reg>,
    ) { ... }
}

Custom handler for system instructions ecall and ebreak

Required Methods

fn handle_ecall( &mut self, regs: &mut Registers<Reg>, memory: &mut Memory, program_counter: &mut PC, ) -> Result<ControlFlow<()>, ExecutionError<Reg::Type, Rv64Instruction<Reg>, CustomError>>

Handle an ecall instruction.

NOTE: the program counter here is the current value, meaning it is already incremented past the instruction itself.

Provided Methods

fn handle_ebreak( &mut self, _regs: &mut Registers<Reg>, _memory: &mut Memory, _pc: Reg::Type, _instruction: Rv64Instruction<Reg>, )

Handle an ebreak instruction.

NOTE: the program counter here is the current value, meaning it is already incremented past the instruction itself.