ab_riscv_interpreter::v

Module vector_registers

Vector registers

Structs

IndexedSegmentElement

Body element of an IndexedSegmentGroup, see IndexedSegmentGroup::elements()

IndexedSegmentGroup

Segment group of an indexed load or store together with its group of indices, see VRegSegmentGroup::with_index()

SegmentElement

Body element of a VRegSegmentGroup, see VRegSegmentGroup::elements()

SegmentField

Field of a SegmentElement, see SegmentElement::fields()

VRegGroup

Register group of elements of width EEW that lies within the vector register file.

VRegSegmentGroup

Register group of a segment load or store: nf fields, each a VRegGroup, in consecutive register groups starting with the first one.

VectorRegisterFile

Vector register file container.

Traits

VectorRegisters

Vector register state.

VectorRegistersExt

Derived convenience accessors for vector CSRs that are simple read/write fields (vstart, vxrm, vxsat, vcsr).