ab_riscv_interpreter::basic

Trait BasicRegister

pub unsafe trait BasicRegister
where
    Self: Register,
{
    const N: usize;

    // Required method
    fn offset(self) -> u8;
}

Basic general purpose register to be used with BasicRegisters

Safety

Self::offset() must return values in 0..Self::N range. Self::from_bits() must return Some() for 0..=31 if Self::RVE = false and 0..=15 if Self::RVE = true.

Required Associated Constants

const N: usize

The number of general purpose registers.

Canonically 32 unless E extension is used, in which case 16.

Required Methods

fn offset(self) -> u8

Offset in a set of registers

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.