ab_riscv_interpreter

Trait ExecutableInstructionCsr

pub trait ExecutableInstructionCsr<ExtState, CustomError = CustomErrorPlaceholder>
where
    Self: Instruction,
    ExtState: ?Sized,
{
    // Provided methods
    fn prepare_csr_read(
        ext_state: &ExtState,
        csr_index: u16,
        raw_value: <<Self as Instruction>::Reg as Register>::Type,
        output_value: &mut <<Self as Instruction>::Reg as Register>::Type,
    ) -> Result<bool, CsrError<CustomError>> { ... }
    fn prepare_csr_write(
        ext_state: &mut ExtState,
        csr_index: u16,
        write_value: <<Self as Instruction>::Reg as Register>::Type,
        output_value: &mut <<Self as Instruction>::Reg as Register>::Type,
    ) -> Result<bool, CsrError<CustomError>> { ... }
}

Provided Methods

fn prepare_csr_read( ext_state: &ExtState, csr_index: u16, raw_value: <<Self as Instruction>::Reg as Register>::Type, output_value: &mut <<Self as Instruction>::Reg as Register>::Type, ) -> Result<bool, CsrError<CustomError>>

Prepare CSR read.

This method is called on each extension one by one with the raw_value (contents of the corresponding CSR register) and initially zero-initialized output_value. In return value every extension can accept (Ok(true)), ignore (Ok(false)) or reject (Err(CsrError)) read request. For accepted reads the extension must update output_value accordingly, which will be the value used by the Zicsr extension handler.

Some extensions will just copy raw_value to output value, others will copy only some bits or zero some bits of the raw_value, as required by the specification.

If no extension returns Ok(true), the read operation is implicitly rejected as illegal access.

fn prepare_csr_write( ext_state: &mut ExtState, csr_index: u16, write_value: <<Self as Instruction>::Reg as Register>::Type, output_value: &mut <<Self as Instruction>::Reg as Register>::Type, ) -> Result<bool, CsrError<CustomError>>

Prepare CSR write.

This method is called on each extension one by one with write_value being prepared by the Zicsr extension handler. In return value every extension can accept (Ok(true)), ignore (Ok(false)) or reject (Err(CsrError)) write request. For accepted writes the extension must update output_value accordingly, which will be written to the corresponding CSR register.

Some extensions will just copy write_value to output value, others will copy some bits or zero some bits of the write_value, as required by the specification.

If no extension returns Ok(true), the write operation is implicitly rejected as illegal access.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.