ab_riscv_interpreter

Trait ProgramCounter

pub trait ProgramCounter<Address, Memory, CustomError = CustomErrorPlaceholder>
where
    Address: Copy,
{
    // Required methods
    fn get_pc(&self) -> Address;
    fn set_pc_relative(
        &mut self,
        memory: &Memory,
        instruction_size: u8,
        offset: i32,
    ) -> Result<ControlFlow<()>, ExecutionError<Address, CustomError>>;
    fn set_pc(
        &mut self,
        memory: &Memory,
        pc: Address,
    ) -> Result<ControlFlow<()>, ExecutionError<Address, CustomError>>;

    // Provided method
    fn old_pc(&self, instruction_size: u8) -> Address
       where Address: From<u8> + Sub<Output = Address> { ... }
}

Generic program counter

Required Methods

fn get_pc(&self) -> Address

Get the current value of the program counter

fn set_pc_relative( &mut self, memory: &Memory, instruction_size: u8, offset: i32, ) -> Result<ControlFlow<()>, ExecutionError<Address, CustomError>>

Apply ExecutionResult::Branch, continuing offset bytes from the instruction being executed.

A simple implementation can resolve the offset against the program counter and call Self::set_pc(). Implementation that keeps the program counter as a position in an already decoded instruction stream can move within that stream directly, which avoids converting an address back into a position.

fn set_pc( &mut self, memory: &Memory, pc: Address, ) -> Result<ControlFlow<()>, ExecutionError<Address, CustomError>>

Set the current value of the program counter

Provided Methods

fn old_pc(&self, instruction_size: u8) -> Address
where Address: From<u8> + Sub<Output = Address>,

Get the previous value of the program counter before executing an instruction.

This is usually called from under instruction execution when the program counter is already advanced during instruction fetching. As such, pc - instruction_size is expected to never underflow.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".